requirements.txt
streamlit==1.15.2
streamlit-aggrid==0.3.3
python-dotenv==0.21.0